% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dti_data.R
\name{delete_dti_data}
\alias{delete_dti_data}
\title{Delete DTI Image Filenames}
\usage{
delete_dti_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "DTI"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the DTI images
}
