% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dti_data.R
\name{download_dti_data}
\alias{download_dti_data}
\title{Download DTI Image Filenames}
\usage{
download_dti_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "DTI"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the DTI images
}
\examples{
download_dti_data()
}
