% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dti_info_filenames.R
\name{get_dti_info_filenames}
\alias{get_b_filenames}
\alias{get_dti_info_filenames}
\alias{get_grad_filenames}
\title{Get DTI Information Filenames}
\usage{
get_dti_info_filenames(ids = kirby21.base::get_ids(), visits = c(1, 2),
  type = c("grad", "b"))

get_grad_filenames(...)

get_b_filenames(...)
}
\arguments{
\item{ids}{ID to return}

\item{visits}{Vector of scan indices to return (1 or 2 or both)}

\item{type}{type of file to obtain (gradient or b-values)}

\item{...}{arguments passed to \code{\link{get_dti_info_filenames}}}

\item{...}{arguments passed to \code{\link{get_dti_info_filenames}}}
}
\value{
Data.frame of filenames
}
\description{
Return the filenames for the files for DTI analysis,
such as gradient files or b-values
}
\examples{
get_dti_info_filenames()
}

