% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dti_filenames.R
\name{get_dti_filenames}
\alias{get_dti_filenames}
\title{Get DTI Image Filenames}
\usage{
get_dti_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "DTI"} so it cannot be specified}
}
\value{
Vector of filenames
}
\description{
Return the filenames for the DTI images
}
\examples{
get_dti_filenames()
}
